--- a/app/views/block/bucket.blade.php
+++ b/app/views/block/bucket.blade.php
@@ -1,11 +1,7 @@
 <?php
 /**
- * @var \Destiny\Character\InventoryBucket $bucket
  * @var \Destiny\Character\InventoryItem $item
  */
-
-$item = $bucket->equipped;
-
 ?>
 
 @if($item)