- Adds /summary to account & inventory endpoints
 - Hacks around Buckets, since each Bucket has 1 item (the equipped one)
 - We don't need to search the buckets for equipped item, so just pass item
 - fixes #12 

BEFORE:
```
destiny/1/account/4611686018430553461/?lc=en (670.09ms)
destiny/1/account/4611686018430553461/character/2305843009240817218/inventory/?lc=en (517.7ms)
destiny/1/account/4611686018430553461/character/2305843009241283261/inventory/?lc=en (515.7ms)
destiny/1/account/4611686018430553461/character/2305843009215679067/inventory/?lc=en (282.61ms)
```

So 670.09 + 517.7 + 515.7 + 282.61 = 1.986s

AFTER:
```
destiny/1/account/4611686018430553461/summary/?lc=en (299.62ms)
destiny/1/account/4611686018430553461/character/2305843009240817218/inventory/summary/?lc=en (242.3ms)
destiny/1/account/4611686018430553461/character/2305843009215679067/inventory/summary/?lc=en (253.01ms)
destiny/1/account/4611686018430553461/character/2305843009241283261/inventory/summary/?lc=en (266.03ms)
```

So 299.62 + 242.3 + 253.01 + 266.03 = 1.060s

So nearly cut the time in half it takes :) This is also on my local computer with some stupid Comcast ISP. Server results should be different. Getting faster and the Bungie team should be happy we aren't using those endpoints anymore.

No visual changes. Site acts just as before, just using different endpoints and leveraging manifest vs old endpoint.